refactor: separate useColorScheme and useRefresh hooks
