+++ src/hooks/useColorScheme.tsx
@@ -1,0 +1,18 @@
+import { useState, useEffect } from "react";
+
+export default function useColorScheme() {
+  const [mode, setMode] = useState<"light" | "dark">("light");
+
+  useEffect(() => {
+    window
+      .matchMedia("(prefers-color-scheme: dark)")
+      .addEventListener("change", (e) => {
+        const colorScheme = e.matches ? "dark" : "light";
+        setMode(colorScheme);
+      });
+  }, []);
+
+  return {
+    colorScheme: mode,
+  };
+}
