+++ src/pages/Home.tsx
@@ -33,7 +33,7 @@ import { chevronForwardOutline } from "ionicons/icons";
 import { getAuth } from "firebase/auth";
 import { useCollectionDataOnce } from "react-firebase-hooks/firestore";
 import useFavorite from "../hooks/useFavorite";
-import { useRefresh } from "../hooks/page";
+import useRefresh from "@/hooks/useRefresh";
 import { useSetRecoilState } from "recoil";
 
 // import ProductCard from "../components/ProductCard";
